Solution to Check MSB of a Number

+++ Check MSB of a Number@@67882a731b4392d8fc1577e9/code.c
@@ -3,7 +3,7 @@
 #include <stdbool.h>
 int main() {
     bool result=False;
-    int a;
+    long a;
     int b=pow(2,31);
     scanf("%d ",&a);
     result= a>=b;
